add rack_init impls
